fix Custom Card List

+++ lib/src/models/current_weather_model.dart
@@ -75,6 +75,6 @@ class Wind {
   Wind({this.speed});
 
   Wind.fromJson(Map<String, dynamic> json) {
-    speed = json['speed'];
+    speed = json['speed'].toDouble();
   }
 }
